WIP: Abstracted vault checkt-setup logic from the checkout flow to be used on the popup-bridge flow as well.
