+++ src/payment-flows/checkout.js
@@ -2,18 +2,19 @@
 
 import { ZalgoPromise } from 'zalgo-promise/src';
 import { memoize, noop, supportsPopups, stringifyError, extendUrl, PopupOpenError } from 'belter/src';
-import { FUNDING, FPTI_KEY, CURRENCY, COUNTRY, INTENT, CARD, type FundingEligibilityType } from '@paypal/sdk-constants/src';
+import { FUNDING, FPTI_KEY } from '@paypal/sdk-constants/src';
 import { getParent, getTop, type CrossDomainWindowType } from 'cross-domain-utils/src';
 
 import type { ProxyWindow, ConnectOptions } from '../types';
 import { type CreateBillingAgreement, type CreateSubscription } from '../props';
-import { enableVault, exchangeAccessTokenForAuthCode, getConnectURL, getFundingEligibility, updateButtonClientConfig, getSmartWallet, loadFraudnet  } from '../api';
+import { exchangeAccessTokenForAuthCode, getConnectURL, updateButtonClientConfig, getSmartWallet, loadFraudnet  } from '../api';
 import { CONTEXT, TARGET_ELEMENT, BUYER_INTENT, FPTI_TRANSITION, FPTI_CONTEXT_TYPE } from '../constants';
 import { unresolvedPromise, getLogger } from '../lib';
 import { openPopup } from '../ui';
 import { FUNDING_SKIP_LOGIN } from '../config';
 
 import type { PaymentFlow, PaymentFlowInstance, SetupOptions, InitOptions } from './types';
+import { enableVaultSetup } from './vault-setup';
 
 export const CHECKOUT_POPUP_DIMENSIONS = {
     WIDTH:  500,
@@ -87,125 +88,6 @@ function isConnectEligible({ connect, vault, fundingSource, createBillingAgreeme
     }
 
     return true;
-}
-
-type IsFundingSourceVaultableOptions = {|
-    accessToken : string,
-    fundingSource : $Values<typeof FUNDING>,
-    clientID : string,
-    merchantID : ?$ReadOnlyArray<string>,
-    currency : $Values<typeof CURRENCY>,
-    buyerCountry : $Values<typeof COUNTRY>,
-    intent : $Values<typeof INTENT>,
-    commit : boolean,
-    vault : boolean,
-    disableFunding : ?$ReadOnlyArray<$Values<typeof FUNDING>>,
-    disableCard : ?$ReadOnlyArray<$Values<typeof CARD>>
-|};
-
-function isFundingSourceVaultable({ accessToken, fundingSource, clientID, merchantID, buyerCountry, currency, commit, vault, intent, disableFunding, disableCard } : IsFundingSourceVaultableOptions) : ZalgoPromise<boolean> {
-    return ZalgoPromise.try(() => {
-        if (fundingSource !== FUNDING.PAYPAL) {
-            return false;
-        }
-
-        return getFundingEligibility(`
-                ${ fundingSource } {
-                    vaultable
-                }
-            `, {
-            accessToken, clientID, merchantID, buyerCountry, currency, commit, vault,
-            intent, disableFunding, disableCard
-        }).then(newFundingEligibility => {
-            if (newFundingEligibility[fundingSource] && newFundingEligibility[fundingSource].vaultable) {
-                return true;
-            } else {
-                return false;
-            }
-        });
-    });
-}
-
-type VaultAutoSetupEligibleProps = {|
-    vault : boolean,
-    clientAccessToken : ?string,
-    createBillingAgreement : ?CreateBillingAgreement,
-    createSubscription : ?CreateSubscription,
-    fundingSource : $Values<typeof FUNDING>,
-    clientID : string,
-    merchantID : ?$ReadOnlyArray<string>,
-    currency : $Values<typeof CURRENCY>,
-    buyerCountry : $Values<typeof COUNTRY>,
-    intent : $Values<typeof INTENT>,
-    commit : boolean,
-    disableFunding : ?$ReadOnlyArray<$Values<typeof FUNDING>>,
-    disableCard : ?$ReadOnlyArray<$Values<typeof CARD>>
-|};
-
-function isVaultAutoSetupEligible({ vault, clientAccessToken, createBillingAgreement, createSubscription, fundingSource,
-    clientID, merchantID, buyerCountry, currency, commit, intent, disableFunding, disableCard } : VaultAutoSetupEligibleProps) : ZalgoPromise<boolean> {
-
-    return ZalgoPromise.try(() => {
-        if (!clientAccessToken) {
-            return false;
-        }
-
-        if (createBillingAgreement || createSubscription) {
-            return false;
-        }
-
-        if (vault) {
-            return true;
-        }
-
-        return isFundingSourceVaultable({ accessToken: clientAccessToken, fundingSource, clientID, merchantID, buyerCountry, currency,
-            commit, vault, intent, disableFunding, disableCard }).catch(err => {
-
-            getLogger().warn('funding_vaultable_error', { err: stringifyError(err) });
-            return false;
-        });
-    });
-}
-
-type EnableVaultSetupOptions = {|
-    orderID : string,
-    vault : boolean,
-    clientAccessToken : ?string,
-    fundingEligibility : FundingEligibilityType,
-    fundingSource : $Values<typeof FUNDING>,
-    createBillingAgreement : ?CreateBillingAgreement,
-    createSubscription : ?CreateSubscription,
-    clientID : string,
-    merchantID : ?$ReadOnlyArray<string>,
-    currency : $Values<typeof CURRENCY>,
-    buyerCountry : $Values<typeof COUNTRY>,
-    intent : $Values<typeof INTENT>,
-    commit : boolean,
-    disableFunding : ?$ReadOnlyArray<$Values<typeof FUNDING>>,
-    disableCard : ?$ReadOnlyArray<$Values<typeof CARD>>,
-    userIDToken : ?string
-|};
-
-function enableVaultSetup({ orderID, vault, clientAccessToken, createBillingAgreement, createSubscription, fundingSource,
-    clientID, merchantID, buyerCountry, currency, commit, intent, disableFunding, disableCard, userIDToken } : EnableVaultSetupOptions) : ZalgoPromise<void> {
-
-    return ZalgoPromise.try(() => {
-        getLogger()
-            .info(`vault_auto_setup_vault_${ vault.toString() }_id_token_${ userIDToken ? 'present' : 'not_present' }`)
-            .flush();
-
-        return isVaultAutoSetupEligible({ vault, clientAccessToken, createBillingAgreement, createSubscription, fundingSource,
-            clientID, merchantID, buyerCountry, currency, commit, intent, disableFunding, disableCard });
-
-    }).then(eligible => {
-        if (eligible && clientAccessToken) {
-            return enableVault({ orderID, clientAccessToken }).catch(err => {
-                if (vault) {
-                    throw err;
-                }
-            });
-        }
-    });
 }
 
 function getContext({ win, isClick } : {| win : ?(CrossDomainWindowType | ProxyWindow), isClick : ?boolean |}) : $Values<typeof CONTEXT> {
@@ -317,6 +199,10 @@ function initCheckout({ props, components, serviceData, payment, config } : Init
             createOrder: () => {
                 return createOrder().then(orderID => {
                     return ZalgoPromise.try(() => {
+                        // eslint-disable-next-line no-console
+                        console.log('vault call --->', clientID && buyerIntent === BUYER_INTENT.PAY, clientID, buyerIntent, BUYER_INTENT.PAY);
+                        // eslint-disable-next-line no-console
+                        console.log('vault prop --->', vault);
                         if (clientID && buyerIntent === BUYER_INTENT.PAY) {
                             return enableVaultSetup({ orderID, vault, clientAccessToken, fundingEligibility, fundingSource, createBillingAgreement, createSubscription,
                                 clientID, merchantID, buyerCountry, currency, commit, intent, disableFunding, disableCard, userIDToken });
